+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## NEXT
+* OpenAI responses API is enabled by default
+* Add new load_default_model function
+* Deprecated load_openai_model function
 
 ## v0.6.4
 * Revert back to use "v" prefix for tags